Add files via upload
I am pleased to share with you an Excel file that serves as an example of the use of various functions and plots in Excel. In addition to the functions, the file also includes visualizations: line plots, bar charts, and pivot charts. These visualizations serve to highlight the relationships between different variables in the data and allow for more in-depth analysis of the data.